Split and rename Uniscribe-related files
This splits `text_drawing.h` and `text_drawing.cpp` into multiple files, more clearly named that they are specifically Uniscribe-based.
