+++ ui_helpers.vcxproj.filters
@@ -17,7 +17,7 @@
     <ClInclude Include="message_hook.h" />
     <ClInclude Include="solid_fill.h" />
     <ClInclude Include="stdafx.h" />
-    <ClInclude Include="text_drawing.h" />
+    <ClInclude Include="uniscribe_text_out.h" />
     <ClInclude Include="win32_helpers.h" />
     <ClInclude Include="ole.h" />
     <ClInclude Include="drag_image.h" />
@@ -48,12 +48,13 @@
     </ClInclude>
     <ClInclude Include="window_subclasser.h" />
     <ClInclude Include="event_token.h" />
+    <ClInclude Include="uniscribe.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="message_hook.cpp" />
     <ClCompile Include="solid_fill.cpp" />
     <ClCompile Include="stdafx.cpp" />
-    <ClCompile Include="text_drawing.cpp" />
+    <ClCompile Include="uniscribe_text_out.cpp" />
     <ClCompile Include="win32_helpers.cpp" />
     <ClCompile Include="ole.cpp" />
     <ClCompile Include="drag_image.cpp" />
